Исправили чтение 64-битных MVA атрибутов; удалили неиспользуемые Sphinx'ом атрибуты
