+++ Sphinx.Client/Commands/Attributes/Values/AttributeValueBase.cs
@@ -65,12 +65,8 @@ namespace Sphinx.Client.Commands.Attributes.Values
             		return new AttributeValueString();
                 case AttributeType.MultiInteger:
                     return new AttributeValuesInt32();
-                case AttributeType.MultiFloat:
-                    return new AttributeValuesFloat();
-                case AttributeType.MultiTimestamp:
-                    return new AttributeValuesDateTime();
-                case AttributeType.MultiBoolean:
-                    return new AttributeValuesBoolean();
+                case AttributeType.MultiLong:
+                    return new AttributeValuesInt64();
             }
             throw new NotSupportedException(String.Format(Messages.Exception_UnsupportedAttributeType, Enum.GetName(typeof(AttributeType), type)));
         }
