+++ Sphinx.Client/Commands/Collections/AttributeUpdateList.cs
@@ -183,97 +183,6 @@ namespace Sphinx.Client.Commands.Collections
 			return item;
 		}
 
-
-        /// <summary>
-        /// Add new boolean multi-value to update attribute value stored in Sphinx index by document ID.
-        /// </summary>
-        /// <param name="name">Attribute name</param>
-        /// <param name="id">Document ID</param>
-        /// <param name="value">New value</param>
-        /// <returns>New object</returns>
-		public AttributeUpdateMultiBoolean Add(string name, long id, IEnumerable<bool> value)
-        {
-			Dictionary<long, IEnumerable<bool>> dict = new Dictionary<long, IEnumerable<bool>>();
-            dict.Add(id, new List<bool>(value));
-            AttributeUpdateMultiBoolean item = new AttributeUpdateMultiBoolean(name, dict);
-            Add(item);
-            return item;
-        }
-
-        /// <summary>
-        /// Add new boolean multi-values to Update real attribute values stored in Sphinx index by document IDs.
-        /// </summary>
-        /// <param name="name">Attribute name</param>
-        /// <param name="values">Document id and override values map to override</param>
-        /// <returns>New object</returns>
-        public AttributeUpdateMultiBoolean Add(string name, IDictionary<long, IEnumerable<bool>> values)
-        {
-            AttributeUpdateMultiBoolean item = new AttributeUpdateMultiBoolean(name, values);
-            Add(item);
-            return item;
-        }
-
-
-        /// <summary>
-        /// Add new DateTime multi-value to update attribute value stored in Sphinx index by document ID.
-        /// </summary>
-        /// <param name="name">Attribute name</param>
-        /// <param name="id">Document ID</param>
-        /// <param name="value">New value</param>
-        /// <returns>New object</returns>
-		public AttributeUpdateMultiDateTime Add(string name, long id, IEnumerable<DateTime> value)
-        {
-			IDictionary<long, IEnumerable<DateTime>> dict = new Dictionary<long, IEnumerable<DateTime>>();
-            dict.Add(id, value);
-            AttributeUpdateMultiDateTime item = new AttributeUpdateMultiDateTime(name, dict);
-            Add(item);
-            return item;
-        }
-
-        /// <summary>
-        /// Add new DateTime multi-values to update attribute values stored in Sphinx index by document IDs.
-        /// </summary>
-        /// <param name="name">Attribute name</param>
-        /// <param name="values">Document id and override values map to override</param>
-        /// <returns>New object</returns>
-		public AttributeUpdateMultiDateTime Add(string name, IDictionary<long, IEnumerable<DateTime>> values)
-        {
-            AttributeUpdateMultiDateTime item = new AttributeUpdateMultiDateTime(name, values);
-            Add(item);
-            return item;
-        }
-
-
-        /// <summary>
-        /// Add new float multi-value to update attribute value stored in Sphinx index by document ID.
-        /// </summary>
-        /// <param name="name">Attribute name</param>
-        /// <param name="id">Document ID</param>
-        /// <param name="value">New value</param>
-        /// <returns>New object</returns>
-		public AttributeUpdateMultiFloat Add(string name, long id, IEnumerable<float> value)
-        {
-			Dictionary<long, IEnumerable<float>> dict = new Dictionary<long, IEnumerable<float>>();
-            dict.Add(id, value);
-            AttributeUpdateMultiFloat item = new AttributeUpdateMultiFloat(name, dict);
-            Add(item);
-            return item;
-        }
-
-        /// <summary>
-        /// Add new float multi-values to update attribute values stored in Sphinx index by document IDs.
-        /// </summary>
-        /// <param name="name">Attribute name</param>
-        /// <param name="values">Document id and override values map to override</param>
-        /// <returns>New object</returns>
-		public AttributeUpdateMultiFloat Add(string name, IDictionary<long, IEnumerable<float>> values)
-        {
-            AttributeUpdateMultiFloat item = new AttributeUpdateMultiFloat(name, values);
-            Add(item);
-            return item;
-        }
-
-
         /// <summary>
         /// Add new int multi-value to update attribute value stored in Sphinx index by document ID.
         /// </summary>
